refactor(context): Rename config mod to context
