+++ src/cargo/core/profiles.rs
@@ -30,7 +30,7 @@ use crate::core::{
 };
 use crate::util::interning::InternedString;
 use crate::util::toml::validate_profile;
-use crate::util::{closest_msg, config, CargoResult, GlobalContext};
+use crate::util::{closest_msg, context, CargoResult, GlobalContext};
 use anyhow::{bail, Context as _};
 use cargo_util_schemas::manifest::TomlTrimPaths;
 use cargo_util_schemas::manifest::TomlTrimPathsValue;
@@ -1298,7 +1298,7 @@ fn merge_config_profiles(
 
 /// Helper for fetching a profile from config.
 fn get_config_profile(ws: &Workspace<'_>, name: &str) -> CargoResult<Option<TomlProfile>> {
-    let profile: Option<config::Value<TomlProfile>> =
+    let profile: Option<context::Value<TomlProfile>> =
         ws.gctx().get(&format!("profile.{}", name))?;
     let Some(profile) = profile else {
         return Ok(None);
