+++ src/cargo/lib.rs
@@ -44,7 +44,7 @@
 //!   directory.
 //! - [`util`]:
 //!   This directory contains generally-useful utility modules.
-//! - [`util::config`]:
+//! - [`util::context`]:
 //!   This directory contains the config parser. It makes heavy use of
 //!   [serde](https://serde.rs/) to merge and translate config values. The
 //!   [`util::GlobalContext`] is usually accessed from the
@@ -117,7 +117,7 @@
 //!     - `src/*/*`: Extracted from `*.crate` by [`sources::registry::RegistrySource`]
 //!   - `git/`: Git source cache.  See [`sources::git`].
 //! - `**/.cargo/config.toml`: Environment dependent (env variables, files) configuration.  See
-//!   [`util::config`]
+//!   [`util::context`]
 //!
 //! ## Contribute to Cargo documentations
 //!
