+++ src/cargo/sources/git/oxide.rs
@@ -301,7 +301,7 @@ pub fn cargo_config_to_gitoxide_overrides(gctx: &GlobalContext) -> CargoResult<V
         Http::USER_AGENT.validated_assignment_fmt(&format!("cargo {}", crate::version()))
     }?);
     if let Some(ssl_version) = &http.ssl_version {
-        use crate::util::config::SslVersionConfig;
+        use crate::util::context::SslVersionConfig;
         match ssl_version {
             SslVersionConfig::Single(version) => {
                 values.push(Http::SSL_VERSION.validated_assignment_fmt(&version)?);
