+++ tests/testsuite/member_discovery.rs
@@ -1,7 +1,7 @@
 //! Tests for workspace member discovery.
 
 use cargo::core::{Shell, Workspace};
-use cargo::util::config::GlobalContext;
+use cargo::util::context::GlobalContext;
 
 use cargo_test_support::install::cargo_home;
 use cargo_test_support::project;
