+++ tests/testsuite/member_errors.rs
@@ -3,7 +3,7 @@
 use cargo::core::resolver::ResolveError;
 use cargo::core::{compiler::CompileMode, Shell, Workspace};
 use cargo::ops::{self, CompileOptions};
-use cargo::util::{config::GlobalContext, errors::ManifestError};
+use cargo::util::{context::GlobalContext, errors::ManifestError};
 
 use cargo_test_support::install::cargo_home;
 use cargo_test_support::project;
